frysk-core/frysk/proc/live/ChangeLog 2007-07-30  Mark Wielaard  <mwielaard@redhat.com>
   * AddressSpaceByteBuffer.java (Request): Make length and offset ints.
   (peek): Likewise.
   * LinuxTask.java (sendrecMemory): Set memory order.
   * LogicalMemoryBuffer.java (order): Remove field.
   (LogicalMemoryBuffer): Don't set order.
   (peek): Make length and offset ints.
   (subBuffer): Don't pass order.
   * MemorySpaceByteBuffer.java (Request): Make length and offset ints.
   (peek): Likewise.
   * TestMemorySpaceByteBuffer.java (AsyncPeeks): Make length int.

frysk-sys/frysk/sys/ChangeLog
2007-07-30  Mark Wielaard  <mwielaard@redhat.com>

   * FileDescriptor.java (read): Make start and length ints.
   (write): Likewise.
   * Ptrace.java (peek): Make length and offset ints.
   * StatelessFile.java (pread): Make start and length ints.
   (pwrite): Likewise.
   * TestFileDescriptor.java (IO.op): Likewise.
   (assertArrayIndexOutOfBounds): Likewise.
   (testReadOutOfBounds): Likewise.
   (testWriteOutOfBounds): Likewise.
   * TestPtrace.java (verifyOutOfBounds): Make length and offset ints.
   * TestStatelessFile.java: Make start and length ints.
   * cni/Errno.cxx (verifyBounds): Make start and length ints.
   * cni/Errno.hxx: Likewise.
   * cni/FileDescriptor.cxx (read): Make start and length jints.
   (write): Likewise.
   * cni/Ptrace.cxx (peek): Make length and offset jints.
   * cni/StatelessFile.cxx (pread): Make start and length jints.
   (pwrite): Likewise.

frysk-sys/inua/ChangeLog
2007-07-30  Mark Wielaard  <mwielaard@redhat.com>

   * eio/ByteBuffer.java (peek): Make off and len ints.
   (poke): Likewise.
   (peekFully): Likewise.
   (pokeFully): Likewise.
   * eio/MmapByteBuffer.java (peek): Likewise.
   * eio/cni/MmapByteBuffer.cxx (peek): Likewise.

+++ frysk-core/frysk/proc/live/AddressSpaceByteBuffer.java
@@ -130,8 +130,8 @@ public class AddressSpaceByteBuffer
 	extends Request
     {
 	private long index;
-	private long length;
-	private long offset;
+	private int length;
+	private int offset;
 	private byte[] bytes;
 	PeeksRequest()
 	{
@@ -141,8 +141,8 @@ public class AddressSpaceByteBuffer
 	{
 	    length = addressSpace.peek(pid, index, length, bytes, offset);
 	}
-	public long request (long index, byte[] bytes,
-			     long offset, long length)
+	public int request (long index, byte[] bytes,
+			    int offset, int length)
 	{
 	    if (isEventLoopThread())
 		return addressSpace.peek(pid, index, length, bytes, offset);
@@ -157,7 +157,7 @@ public class AddressSpaceByteBuffer
 	}
     }
     private final PeeksRequest peeksRequest;
-    protected long peek (long index, byte[] bytes, long offset, long length)
+    protected int peek (long index, byte[] bytes, int offset, int length)
     {
 	return peeksRequest.request(index, bytes, offset, length);
     }
