+++ frysk-sys/frysk/sys/Ptrace.java
@@ -190,8 +190,8 @@ public class Ptrace
 	 * Fetch up-to LENGTH bytes starting at ADDR of process PID,
 	 * store them in BYTES, starting at OFFSET.
 	 */
-	public native long peek (int pid, long addr, long length,
-				 byte[] bytes, long offset);
+	public native int peek (int pid, long addr, int length,
+				byte[] bytes, int offset);
 	private static native AddressSpace text ();
 	private static native AddressSpace data ();
 	private static native AddressSpace usr ();
