+++ frysk-sys/frysk/sys/StatelessFile.java
@@ -63,10 +63,10 @@ public class StatelessFile
 	this(new File(file));
     }
 
-    public native long pread (long fileOffset, byte[] bytes, long start,
-			      long length);
+    public native int pread (long fileOffset, byte[] bytes, int start,
+			     int length);
    
-    public native long pwrite (long fileOffset, byte[] bytes, long start,
-			       long length);
+    public native int pwrite (long fileOffset, byte[] bytes, int start,
+			      int length);
 }
